+++ R/is_rstudio.R
@@ -1,0 +1,3 @@
+is_rstudio <- function(){
+  Sys.getenv("RSTUDIO") == "1"
+}